Add tests for admins and inspectors having access to all scoped contest categories
These tests require no additional code to be written, because contest category access is determined via the contest, all of which admins and inspectors already have access to.

Note the `let!` syntax to enforce the memoized resources to be created. Otherwise the array matchers would pointlessly default to true, since not using the variables in the expected array would lead to their resources not being created in the first place!